requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-multipart==0.0.9
pydantic==2.9.2
google-generativeai==0.7.2
python-dotenv==1.0.1
